Crate ranim

Ranim is an animation engine written in rust based on wgpu, inspired by 3b1b/manim and jkjkil4/JAnim.

Coordinate System

Ranim’s coordinate system is right-handed coordinate:

     +Y
     |
     |
     +----- +X
   /
+Z

Re-exports

pub use ranim_anims as anims;anims

pub use ranim_core as core;

pub use ranim_items as items;items

pub use ranim_render as render;render

pub use core::glam;

pub use inventory;

Modules

cmd

Commands like preview and render

color

Color utilities.

link_magic 🔒

Scene types for dylib / inventory registration and runtime use. Scene types for dylib / inventory registration and runtime use.

prelude

The preludes

scene 🔒

Scene description types (Scene, Output, OutputFormat, etc.) Scene description types.

utils

Utils

Macros

lagged

Construct an AnimLagged with a stagger ratio, pushing each animation in order.

preview_scene

Preview a scene by name.

render_scenerender and non-target_family=wasm

Render a scene by name.

seq

Construct an AnimSequence by playing each animation in order.

stack

Construct an AnimStack by pushing each animation at the same origin.

Structs

Output

The output of a scene

RanimScene

Animation definition builder passed to scene constructors.

SceneConfig

Scene config

Enums

OutputFormat

The output format of a scene

Traits

SceneConstructor

A scene constructor

Functions

find_scene

Return a scene with matched name