ranim::prelude

Trait Color

pub trait Color: FillColor + StrokeColor {
    // Provided method
    fn set_color(&mut self, color: AlphaColor<Srgb>) -> &mut Self { ... }
}

A trait for items that have both fill color and stroke color

This trait is auto implemented for items that implement FillColor and StrokeColor.

Provided Methods

fn set_color(&mut self, color: AlphaColor<Srgb>) -> &mut Self

Setting color(rgba) of an item

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".