ranim

Module utils

Utils

Modules

bezier

Bezier related stuffs

math

Math stuffs

rate_functions

The rate functions

Functions

angle_between_vectors

Get angle between vectors

apart_alpha

Get a target alpha value that can get value of given alpha after mixed n times

avg

Get data’s avg

calculate_hash

Calculate hash

convert_to_2d

Get a 3d point’s 2d coordinate on a 3d plane

convert_to_3d

Get a 2d point’s 3d coordinate on a 3d plane

extend_with_last

Extend the vec with last element

generate_basis

Generate basis vecs for a surface from a unit normal vec

project

Projects a 3D point onto a plane defined by a unit normal vector.

resize_preserving_order

Resize the vec while preserving the order

resize_preserving_order_with_repeated_cnt

Resize the vec while preserving the order

resize_preserving_order_with_repeated_indices

Resize the vec while preserving the order

rotation_between_vectors

Get a rotation matrix from v1 to v2

wrap_point_func_with_point

Apply the function by first transform the points to origin based on a point, then apply the function, then transform the points back.