ranim::utils

Module bezier

Bezier related stuffs

Structs

PathBuilder

A path builder based on quadratic beziers

Functions

approx_cubic_with_quadratic

Approx a cubic bezier with quadratic bezier

cubic_bezier_eval

Evaluate a cubic bezier at t in [0.0, 1.0]

get_subpath_closed_flag

When path is empty, returns None

partial_quadratic_bezier

Returns the control points of the given part of a quadratic bezier curve.

point_on_quadratic_bezier

Returns the point on a quadratic bezier curve at the given parameter.

quad_bezier_eval

Evaluate a quad bezier at t in [0.0, 1.0]

split_cubic_bezier

Split a cubic bezier at progress value t in [0.0, 1.0]

split_quad_bezier

Split a quad bezier at progress value t in [0.0, 1.0]

trim_cubic_bezier

Get [a, b] of a cubic bezier

trim_quad_bezier

Get [a, b] of a quad bezier